CAREER SETU/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
spacy
scikit-learn
pandas
numpy
google-genai
sentence-transformers
pytest
httpx
python-dotenv
pypdf
joblib